requirements.txt
supabase
python-dotenv
python-fasthtml
monsterUi
pydantic
email-validator
